+++ lib/src/events/ready_event.dart
@@ -1,0 +1,5 @@
+import 'package:restrr/src/events/restrr_event.dart';
+
+class ReadyEvent extends RestrrEvent {
+  const ReadyEvent({required super.api});
+}