+++ lib/src/events/request_event.dart
@@ -1,0 +1,8 @@
+import 'package:restrr/restrr.dart';
+
+class RequestEvent extends RestrrEvent {
+  final String route;
+  final int? statusCode;
+
+  const RequestEvent({required super.api, required this.route, this.statusCode});
+}