+++ lib/src/restrr_base.dart
@@ -1,5 +1,6 @@
 import 'package:logging/logging.dart';
 import 'package:restrr/src/cache/batch_cache_view.dart';
+import 'package:restrr/src/events/event_handler.dart';
 import 'package:restrr/src/requests/route.dart';
 import 'package:restrr/src/service/api_service.dart';
 import 'package:restrr/src/service/currency_service.dart';
@@ -10,7 +11,8 @@ import 'cache/cache_view.dart';
 
 class RestrrOptions {
   final bool isWeb;
-  const RestrrOptions({this.isWeb = false});
+  final bool disableLogging;
+  const RestrrOptions({this.isWeb = false, this.disableLogging = false});
 }
 
 enum RestrrInitType { login, register, savedSession }
@@ -18,6 +20,8 @@ enum RestrrInitType { login, register, savedSession }
 /// A builder for creating a new [Restrr] instance.
 /// The [Restrr] instance is created by calling [create].
 class RestrrBuilder {
+  final Map<Type, Function> _eventMap = {};
+
   final RestrrInitType initType;
   final Uri uri;
   String? sessionId;
@@ -37,64 +41,53 @@ class RestrrBuilder {
 
   RestrrBuilder.savedSession({required this.uri}) : initType = RestrrInitType.savedSession;
 
+  RestrrBuilder on<T extends RestrrEvent>(void Function(T) func) {
+    _eventMap[T.runtimeType] = func;
+    return this;
+  }
+
   /// Creates a new session with the given [uri].
   Future<RestResponse<Restrr>> create() async {
-    Restrr.log.info('Attempting to initialize a session (${initType.name}) with $uri');
     // check if the URI is valid
     final RestResponse<HealthResponse> statusResponse = await Restrr.checkUri(uri, isWeb: options.isWeb);
     if (statusResponse.hasError) {
       Restrr.log.warning('Invalid financrr URI: $uri');
-      return statusResponse.error == RestrrError.unknown
-          ? RestrrError.invalidUri.toRestResponse()
-          : statusResponse.error?.toRestResponse(statusCode: statusResponse.statusCode) ??
-              RestrrError.invalidUri.toRestResponse();
+      return (statusResponse.error == RestrrError.unknown
+              ? RestrrError.invalidUri
+              : statusResponse.error ?? RestrrError.invalidUri)
+          .toRestResponse(statusCode: statusResponse.statusCode);
     }
-    Restrr.log.info('Host: $uri, API v${statusResponse.data!.apiVersion}');
+
+    Restrr.log.config('Host: $uri, API v${statusResponse.data!.apiVersion}');
     final RestrrImpl apiImpl = RestrrImpl._(
-        options: options, routeOptions: RouteOptions(hostUri: uri, apiVersion: statusResponse.data!.apiVersion));
-    return switch (initType) {
-      RestrrInitType.register =>
-        _handleRegistration(apiImpl, username!, password!, email: email, displayName: displayName),
-      RestrrInitType.login => _handleLogin(apiImpl, username!, password!),
-      RestrrInitType.savedSession => _handleSavedSession(apiImpl),
+        options: options,
+        routeOptions: RouteOptions(hostUri: uri, apiVersion: statusResponse.data!.apiVersion),
+        eventMap: _eventMap);
+
+    // attempt to authenticate the user
+    final RestResponse<RestrrImpl> apiResponse = await switch (initType) {
+      RestrrInitType.register => _handleAuthProcess(apiImpl,
+          authFunction: () =>
+              apiImpl._userService.register(username!, password!, email: email, displayName: displayName)),
+      RestrrInitType.login =>
+        _handleAuthProcess(apiImpl, authFunction: () => apiImpl._userService.login(username!, password!)),
+      RestrrInitType.savedSession => _handleAuthProcess(apiImpl, authFunction: () => apiImpl._userService.getSelf()),
     };
-  }
 
-  /// Logs in with the given [username] and [password].
-  Future<RestResponse<RestrrImpl>> _handleLogin(RestrrImpl apiImpl, String username, String password) async {
-    final RestResponse<User> response = await apiImpl._userService.login(username, password);
-    if (!response.hasData) {
-      Restrr.log.warning('Invalid credentials for user $username');
-      return RestrrError.invalidCredentials.toRestResponse(statusCode: response.statusCode);
+    // fire [ReadyEvent] if the API is ready
+    if (apiResponse.hasData) {
+      apiImpl.eventHandler.fire(ReadyEvent(api: apiImpl));
     }
-    apiImpl.selfUser = response.data!;
-    Restrr.log.info('Successfully logged in as ${apiImpl.selfUser.username}');
-    return RestResponse(data: apiImpl, statusCode: response.statusCode);
+    return apiResponse;
   }
 
-  /// Registers a new user and logs in.
-  Future<RestResponse<RestrrImpl>> _handleRegistration(RestrrImpl apiImpl, String username, String password,
-      {String? email, String? displayName}) async {
-    final RestResponse<User> response =
-        await apiImpl._userService.register(username, password, email: email, displayName: displayName);
+  Future<RestResponse<RestrrImpl>> _handleAuthProcess(RestrrImpl apiImpl,
+      {required Future<RestResponse<User>> Function() authFunction}) async {
+    final RestResponse<User> response = await authFunction();
     if (response.hasError) {
-      Restrr.log.warning('Failed to register user $username');
       return response.error?.toRestResponse(statusCode: response.statusCode) ?? RestrrError.unknown.toRestResponse();
     }
     apiImpl.selfUser = response.data!;
-    Restrr.log.info('Successfully registered & logged in as ${apiImpl.selfUser.username}');
-    return RestResponse(data: apiImpl, statusCode: response.statusCode);
-  }
-
-  /// Attempts to refresh the session with still saved credentials.
-  Future<RestResponse<RestrrImpl>> _handleSavedSession(RestrrImpl apiImpl) async {
-    final RestResponse<User> response = await apiImpl._userService.getSelf();
-    if (response.hasError) {
-      Restrr.log.warning('Failed to refresh session');
-      return response.error?.toRestResponse(statusCode: response.statusCode) ?? RestrrError.unknown.toRestResponse();
-    }
-    apiImpl.selfUser = response.data!;
-    Restrr.log.info('Successfully refreshed session for ${apiImpl.selfUser.username}');
     return RestResponse(data: apiImpl, statusCode: response.statusCode);
   }
 }
@@ -104,6 +97,8 @@ abstract class Restrr {
 
   /// Getter for the [EntityBuilder] of this [Restrr] instance.
   EntityBuilder get entityBuilder;
+
+  RestrrEventHandler get eventHandler;
 
   RestrrOptions get options;
   RouteOptions get routeOptions;
@@ -119,6 +114,8 @@ abstract class Restrr {
         isWeb: isWeb,
         routeOptions: RouteOptions(hostUri: uri));
   }
+
+  void on<T extends RestrrEvent>(void Function(T) func);
 
   /// Retrieves the currently authenticated user.
   Future<User?> retrieveSelf({bool forceRetrieve = false});
@@ -144,6 +141,9 @@ class RestrrImpl implements Restrr {
   @override
   final RouteOptions routeOptions;
 
+  @override
+  late final RestrrEventHandler eventHandler;
+
   /* Services */
 
   late final UserService _userService = UserService(api: this);
@@ -156,13 +156,17 @@ class RestrrImpl implements Restrr {
 
   late final RestrrEntityBatchCacheView<Currency> _currencyBatchCache = RestrrEntityBatchCacheView();
 
-  RestrrImpl._({required this.options, required this.routeOptions});
+  RestrrImpl._({required this.options, required this.routeOptions, required Map<Type, Function> eventMap})
+      : eventHandler = RestrrEventHandler(eventMap);
 
   @override
   late final EntityBuilder entityBuilder = EntityBuilder(api: this);
 
   @override
   late final User selfUser;
+
+  @override
+  void on<T extends RestrrEvent>(void Function(T) func) => eventHandler.on(func);
 
   @override
   Future<User?> retrieveSelf({bool forceRetrieve = false}) async {
@@ -186,8 +190,8 @@ class RestrrImpl implements Restrr {
   @override
   Future<List<Currency>?> retrieveAllCurrencies({bool forceRetrieve = false}) async {
     return _getOrRetrieveMulti(
-        batchCache: _currencyBatchCache,
-        retrieveFunction: (api) => api._currencyService.retrieveAllCurrencies(),
+      batchCache: _currencyBatchCache,
+      retrieveFunction: (api) => api._currencyService.retrieveAllCurrencies(),
     );
   }